lint-js: Lint app.js
kolibri_app is part of kolibri-tools and as external module this is an
external import so it should be placed just below the router and above
the internals imports. This is different from what's required for
internal plugins because in that case, kolibri_app is part of the
kolibri core project.

This patch also adds an exception for this class to do not require the
use of this on methods.

https://phabricator.endlessm.com/T31232