Add AppView to access to app bundle zip routes
This new view provides a way to access the bundle zip file content
without the need of unzipping it before, all is done in memory.

All apps bundles should be added in the new `apps` directory. This patch
adds a .gitkeep file to ensure that the directory `apps` exists on this
repository.

https://phabricator.endlessm.com/T31250