Proxy custom channel for development
This patch adds a new view that just is a proxy to localhost:8080. This
proxy is used to replace the custom channel presentation zip bundle so
we can preview the result with the custom channel presentation dev
server.

The PROXY_CUSTOM_CHANNEL environ variable should be set on the kolibri
django command environment.

https://phabricator.endlessm.com/T31513