ContentPage: Add button to go back
Ideally this should be a lightbox with a transparent background on top
of the custom presentation. For now add a way to go back in browser
history to the custom presentation home, which is not what we need but
a bit less bad than going back to the channel chooser.

https://phabricator.endlessm.com/T31806