Ensure working environment with pipenv and nodeenv
Add a bootstrap script that automatically ensures all
dependencies. Python dependencies are ensured with Pipenv, while
NodeJS dependencies are ensured with nodeenv.

The node version is set to latest 10x, because upcoming commits will
add lerna which requires a more recent version than the one in current
README. We can't bump to latest node 14 LTS because kolibri-tools
requires node=10x.

Also, update the README and the .gitignore files as needed.

https://phabricator.endlessm.com/T31931