Become a monorepo with yarn workspaces and lerna
Make room to have multiple javascript projects in this repository,
using yarn workspaces and lerna as project manager.

Add a separata package.json for the plugin frontend. Because of the
way Kolibri plugins work, this workspace can't go to the packages/
folder.

Add a lerna configuration file and adapt the package.json with the
same workspaces listed. Also, require yarn in the package.json engine
because that's what we are using.

https://phabricator.endlessm.com/T31931