template: Do not extract css on build
This option makes the styles to be inline in production build. This
change will fix some styles issues that we've detected on IE11 like
arrow buttons not being well positioned.

I don't know exactly the cause of this problem, but different tests have
shown that the styles are applied correctly when inline and with some
errors in other case.

https://phabricator.endlessm.com/T31985