template: Use height instead of min-height for card-body
The flex-grow property doesn't work without a height value in IE11 so
this will fix the problem of Card tags positioning at the bottom.

The change from min-height to height should not have any side effect
because we're truncating all the texts so we should never see any
component overflow.