template: Use input instead of b-form-input in search
This patch replaces the b-form-inupt in the search by a plain
input tag with the corresponding bootstrap class. This change won't
affect the behavior because we aren't using any specific functionality
from b-form-input.

For some reason, the b-form-input was not working correctly in our
build in IE11, so this simple change fixes the problem of not being able
to type in the search field.

https://phabricator.endlessm.com/T32003