Discovery: Add SearchPage
This new page uses the same SearchBar component that's used on custom
channel apps.

The search is done using the kolibri backend, so there's no filtering
operations done on the frontend.

This first implementation just shows the channels that match the query
but not the content cards, that's something that will be done in
following commits.

The search page shows also a list of recommended channels. Right now
these channels are choosen randomly.

https://phabricator.endlessm.com/T31938