components: Move "Cards" components from template-ui
Move a list of components from the template-ui package to the
eos-components to be able to reuse in the Discovery page.

To do this it was needed to remove all the use of vuex in this
components. When it was possible the state was passed as properties and,
in some special cases, the store is used in a generic form, just to use
if it's there, in other case a fallback is used.

The special cases are:
 * getAsset in cardMixin.
 * Card subtitle fallback to state.channel.title when there's no author
   or license_owner.
 * PaginatedCardGrid watch to reset the pagination when the filter
   changes.

A lot of generic functions has been moved from the template-ui store
to the components/utils.js.

This is the list of components moved:
 * Card
 * CardBody
 * CardGrid
 * CarouselCard
 * CollapsibleCardGrid
 * GridPage
 * HighResCard
 * RegularCard
 * LowResCard
 * PaginatedCardGrid
 * SlidableCardGrid
 * mixin/cardMixin

https://phabricator.endlessm.com/T31938