Fix local fonts loading
To make this work a new sass variable was declared,
eos-components-fonts-path, that's used to calculate the real path of the
fonts.

In the template-ui it's easier to fix because we're using vue-cli and we
can modify the way the css is loaded, so in this case we use the module
resolve-url-loader to make it work the relative path.
https://github.com/vuejs/vue-cli/issues/2099#issuecomment-639613478

In the case of the plugin vue modules, I was unable to find the way to
use the resolve-url-loader, but I was able to fix the problem just
setting the variable to the real eos-components fonts path.