Discovery: Fix links to content in search
Add a new state variable customAppContent, to store the content or topic
id and kind to redirect to, inside the custom channel.

For the topic we were using the topic state, but that's not possible for
the content, because the content state is used to store the channel with
the custom channel presentation, so instead of messing around with the
existing state, I've added a new one to and update the code to make it
compatible for both types of content.

https://phabricator.endlessm.com/T31938