component: Adjust the background in ChannelLogo
To improve the visual representation of the ChannelLogo component, we
want to show it at 100% if it's a solid background and at 80% with white
background if it's an image with transparency.

This patch does that using the HTML canvas element, loading the
thumbnail and getting the alpha value of the first pixel.

https://phabricator.endlessm.com/T32072