components: Add new Footer component
This new component is used for the Discovery page Footer and for the
custom presentation.

In the design we've in the custom channel footer the same information
that we've in the ChannelCard so instead of doing the same twice, this
patch adds a variant to the ChannelCard to be able to use it in the
footer with a sightly different style.

This change removes the asset customization for the Footer and now the
'footerImage' asset is not used anymore and will not override the footer
background.

https://phabricator.endlessm.com/T32068