Template: Reimplement filter dropdown as a buttons
Some filters has a lot of options and the plain list is not usable. This
patch changes that design to a list of buttons that takes less vertical
space.

There are also some changes to the app height, to adjust the popover
window behavior, so it doesn't overflow the document and it's well
positioned.

https://phabricator.endlessm.com/T32082