Improve the default theme
Fix headings, buttons, and more in the styles file. Also, add a
mixins.scss to contain a eos-button-outline-variant mixin based on
bootstrap button-outline-variant mixin.

The media type buttons are now part of the theme, so they can be used
directly. Eg: `<b-button variant="video">`.

Because the secondary color changed, and it's the default button
variant in bootstrap-vue:

- Change success colors to secondary.
- Change buttons without variants to light variant.

Also, add a /test page for testing the theme.

https://phabricator.endlessm.com/T31752