scripts: Check the TAG release not the TAG commit
The yarn bump-version command creates a tag with -a so a new object is
created, so the command:

git ls-remote https://github.com/endlessm/kolibri-explore-plugin.git $TAG

Returns the hash of the new tag object and not the head that we want. To
do it correctly we should query the $TAG^{} that resolves to the real
commit pointed by the tag, not to the tag object.

https://phabricator.endlessm.com/T32041