Discovery: Make card grid responsive
This patch makes the discovery page card grid and search grids
responsive, showing three cards for large screens, two for small and
one for really small screens.

A new mixin was introduced to calculate the column width and also to
change the width data in the component on resize, so it can be used to
change inside components properties.

The bootstrap sass breakpoints are also exposed in the styles.scss so it
can be used inside the javascript to determine at witch point to change
the number of elements to show.

https://phabricator.endlessm.com/T32160