template: Add basic "simple bundle" view
Convert the Section view into a dynamic view, similar to the cards. If
the section in the store is a simple bundle, use a custom view. Rename
the previous Section to ListSection.

The custom view is very basic for now. And it is expected that there
will be a PlaylistSection view in the future.

https://phabricator.endlessm.com/T31752