template: Make navbar and header backgrounds match
Add a new XXL breakpoint at 2560 pixels, as defined by Joana for
4K. All header assets will match this width.

The backgrounds now have fixed attachment. So navbar and header
match. The CSS is moved to a mixin for reusing it in the two
components.

The size logic is:
- Screens down to XXL: full height, crop the width.
- Screens up to XXL: full width, crop the height.

The header element now has to have a fixed height for the less-than
XXL to work, because 100% navbar height is only a portion of the 100%
header height.

https://phabricator.endlessm.com/T32175