lightbox: Improve colors and theme properly
It's better to theme bootstrap instead of passing the variants to the
component.

By design none of the theme gray colors are used. A new variable is
added instead for the modal background.

Only the header close button has to be passed as variant because
bootstrap-vue is not picking the right color.

https://phabricator.endlessm.com/T32185