lightbox: Don't scroll and adjust size to breakpoints
The "scrollable" option was slipped in an unrelated task T32137 but it
was never the intention.

The modal itself is already responsive. But the "xl" size is too large
for screens lower than the XL breakpoint. So this uses the responsive
mixin to use the "lg" size in the lower case.

https://phabricator.endlessm.com/T32189