template: Use ContentCard for bundles
The spec for bundles is similar to the content cards, with a play
button.

Ideally we should traverse all nodes and add an isBundle flag at load
time to the topics that are bundles. For now, this is only used in the
test page. Add a FIXME for later, to remove the current way of using
the vuex getter.

Also, update the test page.

https://phabricator.endlessm.com/T32077