template: Ellipsize the header description
And add a button to expand the header and show the description full
when it ellipsizes. Per design, we cut at 5 lines.

Also: when a channel has an image asset in the header, use the primary
color instead of the secondary color as fallback. This makes the
background look better in case Endless curated channels do ellipsize
their description (very unlikely). Add a hasHeaderImage computed
property to the mixin for this.

Also: as per design, use the solid color in the header when the
description is expanded. In other words, set its background image to
none.

https://phabricator.endlessm.com/T32294