components: Move logo inside the margins for header
This patch removes the negative margin for the Endless logo in the
header, so now the logo will be inside the container, like any other
component.

The negative margin is something that doesn't work correctly with
different screen sizes because the margin is not fixed and can be
smaller than the logo size.

With this change we loose the effect of the logo at the left of the
content with the scroll, but if makes it work correctly with diffrent
sizes.

This patch also hides the endless logo for "xs" sizes, showing it only
if there's no breadcrumb.

https://phabricator.endlessm.com/T32279