API: Add fetchChannelAsync method
This new method will allow us to request all the content for a channel
to the backend when there's a lot of content without crashing.

The fetchCollection fails when there's a lot of childs and the hierarchy
is too complex. This new method does the same but instead of querying
all at one time it queries just one level at a time to the backend and
then consolidate the result in the frontend.

https://phabricator.endlessm.com/T32306