discovery page: Add titles to channels page
And move the carousel outside the main container so it gets gray
background as in new specs.

Also add margins exactly as requested in the specs. We can't reach
32px margin with one utility class, so we have to add margin to both
the heading and its container. Also this depends on the spacing of the
neighbour elements.

Also add a v-else for the placeholders and split the carousel and grid
placeholders.

Also: use the $white SASS variable.

https://phabricator.endlessm.com/T32342