discovery page: Feature content only from recommended channels
Add a hardcoded list of the channel IDs that are curated by Endless /
recommended.

Also, get rid of the name convention for asset files in recommended
channels. Instead, use a map to not depend on channel titles that may
change. This also allows having PNG as format, which was requested.

https://phabricator.endlessm.com/T32370