Build: Build template-ui in a separate yarn command
the loading-screen package uses the template-ui default override styles
so it's required that the override folder is the default one. Lerna
builds with multiple threads so to avoid race condition we should ensure
that the template-ui package is built before or after, but not at the
same time.

This patch split the packages build process and does that with two
commands, the first one to build all packages excluding the template-ui
and the second one to build just the template-ui.

https://phabricator.endlessm.com/T32355