template: Pass node to navigation bar
Which in turn passes it to the breadcrumb. Note that the first
breadcrumb won't work in some cases because the channel ID is
hardcoded to "123", so the root topic doesn't match it.

Add an extra check in the breadcrumb because the node may be
undefined.

Also, pass atHome from the home view instead of calculating it in the
component.

https://phabricator.endlessm.com/T32534