template: Remove CSS height rules
Let kolibri hashi take care of the scrolling. Use 100% in all html,
body and #app div to push the footer to the bottom. Avoid vh units
because it doesn't play well with documents inside iframes.

Set overflow hidden in the body because otherwise a scrollbar appears
for a little moment while the hashi resizes the iframe.

https://phabricator.endlessm.com/T32571