template: Add pagination to slidable cardgrid
Add pagination to the slidable cardgrid variant.

At the home view: Initially load 8 cards per section, which means 2
slides for the larger screens / worst case (4 cards per slide).

At the slidable cardgrid: When the carousel reaches the last slide and
there are more cards to load, emit a signal to tell the view that it
needs to load 8 more cards for this section.

Disable the "next" button during the load by using a `loading`
flag. The flag is set to true when the signal is emitted, and is set
back to false when the nodes arrive. For that, the nodes are being
watched.

Also: simplify the `itemsPerSlide` calculation to make it the same as
`itemsPerPage` for larger screens. And use a constant directly instead
of passing the value around.

https://phabricator.endlessm.com/T32607