template: Use pagination in inline sections and bundles
This patch adds the pagination for all the calls to the
getContentByFilter, specifically for the inline sections and bundles.

This implementation fetch all pages for the inline sections and bundles
becuase it "shouldn't" be a lot, but if we detect that this is a problem
in the future, maybe we can do it by demand, adding a button or using
the scrolling event.