template: Don't show resources count on topics
With the new API we don't have all the children in the node object, so
the number is zero. This patch just removes the resources count from the
topic card, because right now it's something that we can't know without
a new request to the server and that could affects to the performance
considerably.