template: Move app overflow-y css to style variable
This will make it configurable for different channel overrides.

The scrollbar is shown always by default to provide a consistent browser
behavior in different views, some with scroll and other without, but for
some channels that uses an iframe (wikipedia) to show the content this
could be a problem because there will be two scrollbars.

https://phabricator.endlessm.com/T32928