template: Refactor and restyle detail views
This component is used by both the bundle section and the content
view. It is mainly a wrapper to provide white background and no
header.

Refactor as per new designs which have the title on top, not side by
side with the thumbnail. The views now pass title and subtitle to the
component as props. This will allow adding a back button in the
component in next commit, vs adding it to each view.

https://phabricator.endlessm.com/T32916