template: Bring back StructuredTags filtering
This patch bring back structured metadata filtering using the default
content tag with an special format (tag=value).

This is done that way because rigth now the new Kolibri structured
metadata is not ready to be used so to not loose this functionality in
the next Endless Key release we can keep using the ContentNode tags for
this.

This custom tag filtering is using the new API, using the tag option,
the only thing done in the frontend is the tag formatting from our
custom tag format (tag=value) to a dropdown button. But the pagination
and everything else is working the same as other filters, just adding
the tag option. So this change doesn't imply any performance issue.

this will be superceeded by the new taxonomy being tested in
https://phabricator.endlessm.com/T32647

This reverts commit a19f2519de77ecab0fecdf071841b498670156c7.

https://phabricator.endlessm.com/T32838