style: Add extra padding to containers
This extra spacing is to allow buttons next to components. Thus the
size is the button size plus the grid gutter.

At the same time, adjust the containers maximun sizes to compensate.

Also add a `.no-container-padding` class to negate this new extra
padding, to allow some components to handle it manually.

For doing this, we have to override the `make-container` Bootstrap
mixin, and apply it again.

This will be used in the row of cards in next commit.

https://phabricator.endlessm.com/T32927