template: Move defaultContentNode logic to a NavigationGuard
Instead doing the redirection in the component, this patch does the
redirection using a vue-router NavigationGuard so this is done before
loading the Home component.

With this change the ability of loading the first ContentNode setting
the defaultContentNode to true is lost, but it's not a big issue for the
Endless Key.

https://phabricator.endlessm.com/T33009