template: Add DetailViewFullScreen component
This new component was in the zim content channel overrides and it's now
in the template components to be able to reuse the same component in
different overrides.

The new option to use this component is showDetailViewFullScreen.

There's also a new option to do not show the full breadcrumb,
showBreadcrumb, that's set to false in all the zim content overrides.

This change make it possible to remove the custom widgets in these
overrides.

https://phabricator.endlessm.com/T33014