template: Keep the search state
Keep alive the Search component to do not loose the search after
clicking on a card and going back.

This patch also introduces a new navigation guard to the Search route so
the search is clean when the from route is the Home. That way the search
is clean when the user clicks in the Search Keywords link.

https://phabricator.endlessm.com/T33025