components: BackToTop, improve design
Like in the slidable cards row. Make it a button with unscaled
Material arrow icon inside.

For making it the same size as the slidable next/previous buttons, we
need to add the 2px border.

Add a shadow to this floating button. Increase the shadow when the
mouse is over the button.

https://phabricator.endlessm.com/T33033