Discovery: Organize channels in categories
Based on a categorization made in Figma. The hardcoded list of
channels is now a categorized list.

Remove the "Explore..." label in the Channels page, also based in new
designs.

Note that this also affects the initial state of the Search page
because we are reusing the component. But this will change in upcoming
commits.

Also note that the design is using a slidable component for these
cards, but that will also happen later.

https://phabricator.endlessm.com/T33092