discovery: Use slidable grid for channels
Refactor slidable grid logic with slots, so it can be used for both
channel cards and content cards.

Use the more generic slidable grid in the landing page, one per
category.

Also, add a new "hasWhiteBackground" property to the slidable grid so
it can work in white backgrounds. This is the case in the landing
page.

https://phabricator.endlessm.com/T33092