Discovery: Drop search results from previous queries
This patch removes the search result from the store to handle it in the
SearchPage component. Instead of storing the search results in the
state, the searchChannels function returns the data to the caller and
the data should be handled there.

This data is not used in any other place, so it could be removed from
the topicsRoot store module and it's stored in the SearchPage component
data directly.

The SearchPage component increments the searchQueryIndex for every
search, so if there's a race condition and a search query done before
ends after a second search query has been launched, it will be dropped
because the searchQueryIndex will be different from the one specified in
the new search query.

https://phabricator.endlessm.com/T33202