Discovery: Integrate zim search with normal results
Integrate the zim search results with the normal search results in the
same location in the app store to render everything as soon as it's
available.

This patch also implements a solution to drop old queries when a new
search query is created, similar to what it's done in the default
search.

https://phabricator.endlessm.com/T33078