InstallContent: Requeue stalled jobs
This patch adds the process ID to the tasks created to import the
channels and requeue this tasks if the process ID is different from the
previous one.

This way the stalled jobs will start again if the application is closed
during a content installation.

The task type is also changed to "REMOTEIMPORT", to show these tasks in
the Kolibri device plugin.

https://phabricator.endlessm.com/T33482