loading: Remove back button from EndlessKey view when required
This patch adds a new path to show the EndlessKey view without the back
button for the returning users that has the database created with the
USB content and the USB is not connected. The USB should be required and
there's no possibility to go back to the welcome page.

https://phabricator.endlessm.com/T33569