Remove axios dependency
In favour of kolibri.client and kolibri.urls for calling the backend.

This makes the communication with the backend consistent with other
Kolibri plugins.

The kolibri.client utility uses axios internally anyways so we don't
have to add the dependency in this project.

Also remove an ApiURL constant. The kolibri.urls utility is the way to
query URLs by name.

https://phabricator.endlessm.com/T33835