Add CollectionDownloadManager
It is a singleton object that will manage the download of an EK
collection manifest from the network. It will behave the same as doing
the following from the command line:

```
kolibri manage importchannel network $CHANNEL_ID
kolibri manage importcontent --manifest=$PATH_TO_MANIFEST network
$CHANNEL_ID
```

For each channel in the manifest.

It queues one job at a time and won't continue until the job is
completed. In case the job fails or gets cancelled the job will be
restarted, assuming it will be completed after a number of
retries. There is a FIXME note to improve this by giving up after a
number of failures.

The download is done in two phases: 1. channels are imported first so
we get their metadata for the next phase. The UI can then display
something like "importing channels metadata (1 of 13)…". And 2. the
content is imported.

With this granularity the UI can display detailed information of
what's going on to the user like "importing TED-Ed… 25%" and even
display the channel icon.

The download state is serialized and saved in the request session,
which Kolibri stores in encrypted files inside
`KOLIBRI_HOME/sessions/` by default. So a previous download can be
resumed.

Use the CollectionDownloadManager to implement the remaining API:
- `GET get_should_resume`
- `POST start_download [grade, name]`
- `POST resume_download`
- `POST update_download`
- `DELETE cancel_download`
- `GET get_download_status`

https://phabricator.endlessm.com/T33835