Welcome screen: Add modal with privacy policy
In the welcome / get started view. Add a link that pops up a modal with
the text.

The modal is themed in the same way as other modals using the awkward
deep selector for vue SASS.

Note that we are not marking labels as translatable in the welcome
screen yet. This is a standalone webapp, not a kolibri plugin.

https://phabricator.endlessm.com/T33905