Discovery: Implement collections landing page
It has sections of different tagged content. And a "This week's channel"
which is placeholder for now.

The sections are left explicit in the view template because the channel
section above is a especial case and it's in the middle of the "normal"
sections. But the sections tags are defined a an array in the constants
module and are fetched all at once from the backend.

In case a section has missing nodes it won't be present in the UI.

https://phabricator.endlessm.com/T34285