Search page: Improve parsing nodes
The parseNodes isn't used anymore by the template, that leaves the
search page as the only user. And the search page doesn't use the
isBundle parameter at all, so it can map the nodes with structured tags
directly.

Also, it turns out that adding the thumbnail isn't needed anymore. And
tweaking the URL for the Discovery page can happen in a separate mapping
function.

The same tweaks will come handy in the new collections view.

https://phabricator.endlessm.com/T34285