Add PackCard component
The pack ID, title and subtitle props will be passed by the view reading
the pack metadata from the previous commit.

The card has a selected state that can be set passing the selected
boolean prop. When selected it displays more information and the
"choose" button.

Clicking the "choose" button emits the choosePack signal so the parent
view can hook to it.

Please note that visuals and images are placeholders.

https://phabricator.endlessm.com/T34380