Rename previous landing page and default to packs
Rename previous components to "EK Iguana". Default to the new packs
landing page.

Do the renaming of views and components. And reuse the feature flag to
be USE_EK_IGUANA_PAGE. When this setting is enabled, the old hardcoded
list of channels is shown as sections in the landing page. Otherwise,
the new section of tagged content is shown.

https://phabricator.endlessm.com/T34363