Collections backend: Attempt all packs manifests
Try to load all the manifest files for the initial/starter packs. If
they aren't found print an error log instead of raising an exception
that would crash the backend.

Also: don't validate the packs for now because the editor is not ready
to do that yet. Add a TODO note about this.

https://phabricator.endlessm.com/T34363