Packs download: Handle insufficient space
Backend changes: Handle the job error and raise a new exception with
translated message for the UI. This prevents the error to be retried
over and over again.

Frontend changes: Catch the error. Add a new errorMessage data to the
component. When the error is catched: copy its message, stop the update
loop, and display it in the UI along with a Retry button.

https://phabricator.endlessm.com/T34520