Add new download page
The / URL redirects either to the download page (if conditions are met)
or to the landing page. For this there is a new decideDownload handler.

The new DownloadPage view mimics the WelcomeBase from the Welcome
webapp. That is, all content centered. Ideally WelcomeBase should be
moved to eos-components but since the design specs are being updated, I
prevented this change.

All logic inside the DownloadPage is copied from the
InstallContentModal, which will be removed in next commit.

https://phabricator.endlessm.com/T34509