About modal: Add cheat code to Device tab
Allow the modal main div to receive keyboard input. Add a listener to
the keyup event. Add to check when the user completed the cheat code and
to reset the state when a different character is pressed. Load the
Device tab in same window when the user completes the cheat code.

https://phabricator.endlessm.com/T34797