Use Learn content viewer when available
If the Learn plugin is present, build a URL with the content ID and
navigate to it. Also pass the current location as URL parameter. If the
Learn plugin can handle that parameter, it should use it to navigate
back.

If the Learn plugin is not present (no URL matching) display the content
in a lightbox modal as before.

https://phabricator.endlessm.com/T34740