CI: Add bump2release action to bump the version and tag for release
Add bump2release action to bump the version and tag for release. This is
triggered by dispatch with MAJOR or MINOR version bumping.

This action will push a new commit bumping the version and a new
corresponding tag at the end. That will trigger the main action to build
and release automatically.

https://phabricator.endlessm.com/T34804