EkCardBody: Rearrange rows
The goal of this commit is to move the channel icon and name
to above the node title, but doing so wasn't as straightforward
as shuffling code around.

Simply moving the current elements make the title of the node
be bottom aligned, which is undesired. To make it look as the
mockups, add a v-layout element with 3 rows; the first row for
the channel icon and title, the second for the node title, and
the third for the description.

The logic to show either the channel name or the description
has been preserved.