Download: Request all thumbnails for channel
In order to provide a rich display of unavailable content, thumbnails
for the missing content are desired. After the initial starter pack
download has completed, queue up tasks to download all missing content
thumbnails in the background. This uses a new `all_thumbnails` option
for the `remotecontentimport` Kolibri task added after 0.16.0-alpha14.
The option is ignored on older releases.

This introduces a new `FOREGROUND_COMPLETED` stage in for the download
manager. It has no tasks in it, but it separates the foreground tasks
from the background tasks. Once `FOREGROUND_COMPLETED` has been reached,
progress is reported to the frontend as complete and the remaining tasks
will be enqueued in the background without any interaction from the
frontend.

Fixes: #549