CI: Fix checkout branch for bump version from wrokflow
The actions/checkout@v3's ref [1] is the reference or SHA when trigger
the workflow by default. So, the checkout action will not checkout at
the tag until the "ref" is specified with a tag name.

This commit lists the scenarios of both workflow_call and
non-workflow_call for actions/checkout and softprops/action-gh-release.

[1]: https://github.com/actions/checkout/blob/v3.5.3/README.md#usage

Fixes: https://github.com/endlessm/kolibri-explore-plugin/issues/589