Downloadable content card UI and states
Note: Up to this point, the card download is only a mock in the
template-ui test page.

The card relies in a Vue plugin that exposes a global instance
property `$download` with methods for starting, checking and retrying
downloads.

Unavailable content cards will first check if there is a download
state on mounted. Depending on the state they will show a download
button (when ready), a check mark (when completed) or a retry button
(when failed).

When downloading, the card will poll the check method for updating
itself.

The card also emits a "node updated" signal when completed. This is
passed through EkCard as well.