Add new manageContent plugin submodule
This is a Vuex module that stores the content download tasks indexed
by channel ID and content ID.

There are mutations to set all at once (not yet used) and to set the
download task for a particular channel and content IDs.

There is a getter to obtain the task, given the channel and content
IDs.

Finally, there are handlers to implement the content download start,
check, and retry. In there the Kolibri task status is mapped to a
download state that the UI understands.