Rename 'showUnavailable' to 'hideUnavailable'
We are going to invert the logic of the UI element, so this commit
renames the respective variables to make this easier to read.

It also renames 'NO_AVAILABLE_FILTERING' to 'DEFAULT_HIDE_UNAVAILABLE',
to reflect that constant's true meaning now that we removed the option
to enable or disable the "show unavailable" feature.

This commit brings no functional changes.

https://github.com/endlessm/kolibri-explore-plugin/issues/596