Home: Add check to hide unavailable content
Add a checkbox that is effectively equal to the
one added to the Channel's search view.

Changing the switch fetches the list of nodes again,
passing the appropriate flag to KolibriApi. This is
not optimal. One potential improvement is to always
fetch unavailable content, and filter it within the
view, but I've gone with this simpler approach given
the time constraints.