Add background job queue and restart failed jobs during startup
In order to differentiate foreground tasks from background tasks, a
separate background job queue is added. When the plugin is loaded, it
will immediately restart any failed background tasks. No frontend
integration is added here, but the idea is that the frontend will query
both the foreground and background queues for completion and update the
UI as needed.