Download: Use correct channel name in import tasks
We don't currently use this in the frontend, but we might and there's no
reason not to put the correct information there. At a minimum, it makes
the logs more useful. While here, drop the `extra_metadata` field in the
content import tasks as it's ignored by the Kolibri tasks system.