kolibriApi: Fetch node disabling cache
Add an option to fetch the node by removing it from the cache, if it's
cached.

In general this wouldn't be needed but the custom presentation is
behaving differently and keeps the content cached, even when the backed
is correctly setting a new cache key by calling
`ContentCacheKey.update_cache_key()`.