footer: Add app install links for Play and Microsoft stores
These will allow the user to install the Endless Key app, if that would
be a better fit for them on their device than using key.endlessos.org in
a browser.

The badges came from:
 - https://play.google.com/intl/en_us/badges/
 - https://apps.microsoft.com/store/app-badge

Getting them the same size was a complete pain because, for no good
reason, Google have baked a transparent border into the image they serve
to you. We have to use the image they serve because otherwise we’d have
to carry 20 different localised versions.

Changing the layout to match the design
(https://github.com/endlessm/endless-key-content-private/issues/64#issuecomment-1592442043)
required splitting the footer bar into two separate containers, to allow
for grouping the links/buttons at the start and end of the footer row.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://github.com/endlessm/endless-key-content-private/issues/64