Collections page: try fetching nodes on mounted
There is a chance that the core loading state is already set to false
(already loaded). So watching the state for fetching the highlighted
content will never run.

Thus, we need to check the loading state when the view is mounted too.

This is the correct solution to the previous workaround added in #598