footer: Use Native App Install Prompt to install Android app
When supported, use this API to provide a slightly more streamlined
experience for installing the Android app when pressing the “Get it on
Google Play” button.

The API is only supported by Chrome, on mobile, if the app is advertised
as a related application in the PWA manifest for the site, and if the
app is not already installed.

The full list of requirements are here:
https://developer.chrome.com/blog/app-install-banners-native/#what-are-the-criteria

If the API is not supported, the button will function as a link as
normal.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://github.com/endlessm/endless-key-content-private/issues/64